+++ src/aui-form-builder/js/aui-form-builder-field-toolbar.js
@@ -54,7 +54,8 @@ A.FormBuilderFieldToolbar = A.Base.create('form-builder-field-toolbar', A.Base, 
             this._eventHandles.push(
                 fieldsContainer.delegate('click', this._onFieldClick, '.' + CSS_FIELD_CONTENT, this)
             );
-        } else {
+        }
+        else {
             this._eventHandles.push(
                 fieldsContainer.delegate('mouseenter', this._onFieldMouseEnter, '.' + CSS_FIELD_CONTENT_TOOLBAR, this),
                 fieldsContainer.delegate('mouseleave', this._onFieldMouseLeave, '.' + CSS_FIELD_CONTENT_TOOLBAR, this)
@@ -86,7 +87,7 @@ A.FormBuilderFieldToolbar = A.Base.create('form-builder-field-toolbar', A.Base, 
     },
 
     /**
-     * Closes the toolbar
+     * Closes the toolbar.
      *
      * @method close
      */
@@ -95,10 +96,11 @@ A.FormBuilderFieldToolbar = A.Base.create('form-builder-field-toolbar', A.Base, 
     },
 
     /**
-     * Finds one toolbar's item using the selector
+     * Finds one toolbar's item using the selector.
      *
      * @method getItem
      * @param {String} selector
+     * @return {Node}
      */
     getItem: function(selector) {
         return this._toolbar.one(selector);