+++ src/aui-form-builder/HISTORY.md
@@ -4,6 +4,7 @@
 
 ## @VERSION@
 
+* [AUI-1928](https://issues.liferay.com/browse/AUI-1928) Fields should be moved between others field in a same col
 * [AUI-1926](https://issues.liferay.com/browse/AUI-1926) Change functionality of move rows
 * [AUI-1918](https://issues.liferay.com/browse/AUI-1918) Change functionality of remove rows
 * [AUI-1917](https://issues.liferay.com/browse/AUI-1917) Add option to create Fields between two others Fields on the same column